# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TechUniverse",
        "description": "Documentacion de la API TechUniverse",
        "version": "1.2.0"
    },
    "servers": [
        {
            "url": "https://ivan123.alwaysdata.net",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/especificacion/productos": {
            "get": {
                "tags": [
                    "Especificaciones"
                ],
                "summary": "Listar todas las especificaciones con sus productos",
                "description": "Retorna una lista de todas las especificaciones incluyendo sus productos asociados",
                "operationId": "05efd52056bc1480985d5dae08ecbdea",
                "responses": {
                    "200": {
                        "description": "Lista de especificaciones con productos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EspecificacionConProductos"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Producto Especificaciones"
                ],
                "summary": "Añadir una especificacion a un producto",
                "description": "Asocia una especificacion con un valor a un producto",
                "operationId": "b7d29292eea2572c57c3b814db7a7048",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "producto_id",
                                    "especificacion_id",
                                    "valor"
                                ],
                                "properties": {
                                    "producto_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "especificacion_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "valor": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "16GB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Especificacion añadida al producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Especificacion añadida al producto"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductoEspecificacion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validacion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The producto_id field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/cesta/productos": {
            "post": {
                "tags": [
                    "Cesta Productos"
                ],
                "summary": "Añadir un producto a la cesta",
                "description": "Añade un producto a la cesta del usuario autenticado. Si ya existe, incrementa la cantidad",
                "operationId": "69ca7d62ece9ab43223c37faa2d161ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "producto_id",
                                    "cantidad"
                                ],
                                "properties": {
                                    "producto_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto añadido a la cesta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto añadido a la cesta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cesta no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Cesta no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validacion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The producto_id field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cesta/productos/{id}": {
            "put": {
                "tags": [
                    "Cesta Productos"
                ],
                "summary": "Actualizar la cantidad de un producto en la cesta",
                "description": "Actualiza la cantidad de un producto especifico en la cesta del usuario autenticado",
                "operationId": "68b4ed48c51baceae0ff5f38562b5d75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad"
                                ],
                                "properties": {
                                    "cantidad": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cantidad actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cantidad actualizada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado en la cesta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Producto no encontrado en la cesta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cesta Productos"
                ],
                "summary": "Eliminar un producto de la cesta",
                "description": "Elimina un producto especifico de la cesta del usuario autenticado",
                "operationId": "8b3df6e92cfa39b9e3177f80a3a1d65e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar de la cesta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado de la cesta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado de la cesta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado en la cesta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Producto no encontrado en la cesta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos con paginación",
                "description": "Retorna una lista paginada de productos, incluyendo su categoría y especificaciones. Permite ordenación opcional por precio o novedad.",
                "operationId": "70a301af8ea04ce2f2596a58758632cd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página a obtener",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Criterio de ordenación opcional",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "precio_asc",
                                "precio_desc",
                                "novedad_asc",
                                "novedad_desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de productos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 60
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un producto nuevo con imagen subida a R2",
                "operationId": "4ff8851cb6bfe92aa226cdc804917ff0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "categoria_id",
                                    "foto"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Laptop Gaming"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 1500
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Descripción opcional",
                                        "nullable": true
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2
                                    },
                                    "foto": {
                                        "description": "Imagen del producto (máx. 4MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Producto creado con éxito"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The nombre field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Especificacion": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "RAM"
                    }
                },
                "type": "object"
            },
            "EspecificacionConProductos": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "RAM"
                    },
                    "productos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Laptop Gaming"
                                },
                                "valor": {
                                    "type": "string",
                                    "example": "16GB"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductoEspecificacion": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "producto_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "especificacion_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "valor": {
                        "type": "string",
                        "example": "16GB"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "especificacion": {
                        "$ref": "#/components/schemas/Especificacion"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}